requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-django==4.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.0
